Allow codepage to be specified when session is created; default to UTF-8 encoding
For this change, replace cStringIO with StringIO to handle Unicode strings. Ensure we use cp437 decoding when the console is set to codepage 437.
Also add response text when an unrecognized HTTP error occurs communicating with the Windows server. Don't print error message to console if XML was invalid, but return silently.